Refactor code
- move code
- reduce type visibility
- add some comments
