+++ src/test/mod.rs
@@ -9,9 +9,9 @@ mod cancel_opening;
 mod deactivate_application;
 mod ensure_can_add_application;
 mod fill_opening;
+mod on_finalize;
 mod smoke;
 mod unstaked;
-mod on_finalize;
 
 use super::*;
 use crate::mock::Test;
@@ -22,7 +22,6 @@ pub(crate) type BlockNumber = <Test as system::Trait>::BlockNumber;
 pub(crate) type StakeId = <Test as stake::Trait>::StakeId;
 pub(crate) type Balance =
     <<Test as stake::Trait>::Currency as Currency<<Test as system::Trait>::AccountId>>::Balance;
-
 
 //Debug test object content
 pub(crate) fn debug_print<T: rstd::fmt::Debug>(obj: T) {
